Move code close to caller
