+++ internal/cmd/workflow/list.go
@@ -1,7 +1,9 @@
 package workflow
 
 import (
+	"encoding/json"
 	"fmt"
+	"time"
 
 	"github.com/planetscale/cli/internal/cmdutil"
 	"github.com/planetscale/cli/internal/printer"
@@ -49,4 +51,57 @@ func ListCmd(ch *cmdutil.Helper) *cobra.Command {
 	}
 
 	return cmd
+}
+
+type MinimalWorkflow struct {
+	Number      uint64 `header:"Number"`
+	Name        string `header:"name"`
+	State       string `header:"state"`
+	CreatedAt   int64  `header:"created_at,timestamp(ms|utc|human)" json:"created_at"`
+	CompletedAt *int64 `header:"completed_at,timestamp(ms|utc|human)" json:"completed_at"`
+	Duration    int64  `header:"duration,unixduration" json:"duration"`
+
+	orig *ps.Workflow
+}
+
+func toMinimalWorkflows(workflows []*ps.Workflow) []*MinimalWorkflow {
+	minimalWorkflows := make([]*MinimalWorkflow, 0, len(workflows))
+
+	for _, w := range workflows {
+		minimalWorkflows = append(minimalWorkflows, toMinimalWorkflow(w))
+	}
+
+	return minimalWorkflows
+}
+
+func toMinimalWorkflow(w *ps.Workflow) *MinimalWorkflow {
+	duration := durationIfExists(&w.CreatedAt, w.CompletedAt)
+
+	return &MinimalWorkflow{
+		Number:      w.Number,
+		Name:        w.Name,
+		State:       w.State,
+		CreatedAt:   printer.GetMilliseconds(w.CreatedAt),
+		CompletedAt: printer.GetMillisecondsIfExists(w.CompletedAt),
+		Duration:    duration.Milliseconds(),
+		orig:        w,
+	}
+}
+
+func durationIfExists(start *time.Time, end *time.Time) time.Duration {
+	var duration time.Duration
+
+	if start != nil && end != nil {
+		duration = end.Sub(*start)
+	}
+
+	return duration
+}
+
+func (w *MinimalWorkflow) MarshalJSON() ([]byte, error) {
+	return json.MarshalIndent(w.orig, "", " ")
+}
+
+func (w *MinimalWorkflow) MarshalCSVValue() interface{} {
+	return []*MinimalWorkflow{w}
 }
