+++ internal/cmd/workflow/workflow.go
@@ -1,12 +1,7 @@
 package workflow
 
 import (
-	"encoding/json"
-	"time"
-
 	"github.com/planetscale/cli/internal/cmdutil"
-	"github.com/planetscale/cli/internal/printer"
-	ps "github.com/planetscale/planetscale-go/planetscale"
 	"github.com/spf13/cobra"
 )
 
@@ -25,57 +20,4 @@ func WorkflowCmd(ch *cmdutil.Helper) *cobra.Command {
 	cmd.AddCommand(ShowCmd(ch))
 
 	return cmd
-}
-
-type MinimalWorkflow struct {
-	Number      uint64 `header:"Number"`
-	Name        string `header:"name"`
-	State       string `header:"state"`
-	CreatedAt   int64  `header:"created_at,timestamp(ms|utc|human)" json:"created_at"`
-	CompletedAt *int64 `header:"completed_at,timestamp(ms|utc|human)" json:"completed_at"`
-	Duration    int64  `header:"duration,unixduration" json:"duration"`
-
-	orig *ps.Workflow
-}
-
-func toMinimalWorkflows(workflows []*ps.Workflow) []*MinimalWorkflow {
-	minimalWorkflows := make([]*MinimalWorkflow, 0, len(workflows))
-
-	for _, w := range workflows {
-		minimalWorkflows = append(minimalWorkflows, toMinimalWorkflow(w))
-	}
-
-	return minimalWorkflows
-}
-
-func toMinimalWorkflow(w *ps.Workflow) *MinimalWorkflow {
-	duration := durationIfExists(&w.CreatedAt, w.CompletedAt)
-
-	return &MinimalWorkflow{
-		Number:      w.Number,
-		Name:        w.Name,
-		State:       w.State,
-		CreatedAt:   printer.GetMilliseconds(w.CreatedAt),
-		CompletedAt: printer.GetMillisecondsIfExists(w.CompletedAt),
-		Duration:    duration.Milliseconds(),
-		orig:        w,
-	}
-}
-
-func durationIfExists(start *time.Time, end *time.Time) time.Duration {
-	var duration time.Duration
-
-	if start != nil && end != nil {
-		duration = end.Sub(*start)
-	}
-
-	return duration
-}
-
-func (w *MinimalWorkflow) MarshalJSON() ([]byte, error) {
-	return json.MarshalIndent(w.orig, "", " ")
-}
-
-func (w *MinimalWorkflow) MarshalCSVValue() interface{} {
-	return []*MinimalWorkflow{w}
 }
